Trying to clean shit.
