+++ settings.py
@@ -4,10 +4,8 @@
 # download_folder = "./Downloads/" - will download in the CyberDropDownloader folder
 # download_attempts = 10 - will try and download each file 10 times before giving up.
 # include_id_in_download_folder_name = False - setting this to true will add the ID to download folder names.
-# post_scraping_readout = False - setting this to true will output every scraped link that'll be used for downloads
 
 threads = 0
 download_folder = "./Downloads/"
 download_attempts = 10
 include_id_in_download_folder_name = False
-post_scraping_readout = False
